requirements.txt
# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
torch>=1.7.0
torchvision
tqdm>=4.64.0
protobuf<=3.21.6

# Plotting ------------------------------------
pandas>=1.1.4
seaborn

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow